+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## git-lost
 
+## 0.2.0
+
+- Added action parameter to allow for different commands: dirty, branch and list
+- Ahead or behind branches showing up in dirty
+- Anything off the default branch showing up in branch
+- List shows all repositories found
+
 ## 0.1.0
 
 - initial release :tada: